Move case display wrapper test out of API tests
